chore: refactor internal and create external form index pages
